Self management of c_if gates in all simulators. Additionaly, the executors of each simulator are in a single .hpp to avoid code repeating
