+++ cunqa/qpu.py
@@ -6,6 +6,8 @@ import os
 from typing import  Union, Any, Optional
 import inspect
 
+from qiskit import QuantumCircuit
+
 from cunqa.qclient import QClient
 from cunqa.circuit import CunqaCircuit 
 from cunqa.backend import Backend
@@ -72,7 +74,7 @@ class QPU:
     def backend(self) -> Backend:
         return self._backend
 
-    def run(self, circuit: Union[dict, 'CunqaCircuit'], transpile: bool = False, initial_layout: Optional["list[int]"] = None, opt_level: int = 1, **run_parameters: Any) -> 'QJob':
+    def run(self, circuit: Union[dict, 'CunqaCircuit', 'QuantumCircuit'], transpile: bool = False, initial_layout: Optional["list[int]"] = None, opt_level: int = 1, **run_parameters: Any) -> 'QJob':
         """
         Class method to run a circuit in the QPU.
 
@@ -100,7 +102,7 @@ class QPU:
         <class 'QJob'> object.
         """
         # Disallow execution of distributed circuits
-        if inspect.stack()[1].function != "run_distributed":
+        if inspect.stack()[1].function != "run_distributed": # Checks if the run() is called from run_distributed()
             if isinstance(circuit, CunqaCircuit):
                 if circuit.is_distributed:
                     logger.error("Distributed circuits can't run using QPU.run(), try run_distributed() instead.")
@@ -110,7 +112,7 @@ class QPU:
                     logger.error("Distributed circuits can't run using QPU.run(), try run_distributed() instead.")
                     raise SystemExit
             else:
-                logger.error(f"Circuit must be <class cunqa.circuit.CunqaCircuit> or dict, but {type(circuit)} was provided [{TypeError.__name__}]")
+                logger.error(f"Circuit must be a dict, a <class cunqa.circuit.CunqaCircuit> or a <class 'qiskit.circuit.quantumcircuit.QuantumCircuit'>, but {type(circuit)} was provided [{TypeError.__name__}]")
                 raise SystemExit
