+++ examples/ccomm_examples/distributed_raw.py
@@ -37,12 +37,14 @@ d_qc_0_zmq = {
     {
         "name": "measure",
         "qubits": [0],
-        "clbits":[0]
+        "clbits":[0], 
+        "clreg":[]
     },
     {
         "name": "measure",
         "qubits": [1],
-        "clbits": [1]
+        "clbits": [1],
+        "clreg":[]
     }
     ],
     "is_distributed": True,
@@ -73,12 +75,14 @@ d_qc_1_zmq = {
     {
         "name": "measure",
         "qubits": [0],
-        "clbits": [0]
+        "clbits": [0], 
+        "clreg":[]
     },
     {
         "name": "measure",
         "qubits": [1],
-        "clbits": [1]
+        "clbits": [1], 
+        "clreg":[]
     }
     ],
     "is_distributed": True,
@@ -109,12 +113,14 @@ d_qc_2_zmq = {
     {
         "name": "measure",
         "qubits": [0],
-        "clbits": [0]
+        "clbits": [0], 
+        "clreg":[]
     },
     {
         "name": "measure",
         "qubits": [1],
-        "clbits": [1]
+        "clbits": [1], 
+        "clreg":[]
     }
     ],
     "is_distributed": True,
@@ -138,7 +144,7 @@ d_qc_2_zmq = {
 
 from cunqa.mappers import run_distributed
 
-job0, job1, job2 = run_distributed([d_qc_0_zmq, d_qc_1_zmq, d_qc_2_zmq], qpus[:3], shots = 1)
+job0, job1, job2 = run_distributed([d_qc_0_zmq, d_qc_1_zmq, d_qc_2_zmq], qpus[:3], shots = 10)
 
 print("Result QPU0", job0.result.counts)
 print("Result QPU1", job1.result.counts)
