+++ src/backends/simulators/CUNQA/cunqa_classical_comm_simulator.cpp
@@ -1,142 +1,15 @@
 #include "cunqa_classical_comm_simulator.hpp"
-
-#include "src/instructions.hpp"
-#include "src/result_cunqasim.hpp"
-#include "src/executor.hpp"
-#include "src/utils/types_cunqasim.hpp"
+#include "cunqa_executors.hpp"
 
 #include "utils/json.hpp"
-#include "utils/constants.hpp"
-#include "logger.hpp"
 
 namespace cunqa {
 namespace sim {
 
 JSON CunqaCCSimulator::execute(const ClassicalCommBackend& backend, const QuantumTask& quantum_task)
 {
-    std::vector<std::string> connect_with = quantum_task.sending_to;
-    this->classical_channel->set_classical_connections(connect_with);
-
-    std::vector<JSON> instructions = quantum_task.circuit;
-    JSON run_config = quantum_task.config;
-    int shots = run_config.at("shots");
-    std::string instruction_name;
-    std::vector<int> qubits;
-    std::vector<std::string> endpoint;
-    std::vector<double> param;
-    Matrix matrix;
-    int measurement;
-    JSON result;
-    std::unordered_map<int, int> counts;
-    float time_taken;
-
-    int n_qubits = quantum_task.config.at("num_qubits");
-    if (executor != nullptr && (n_qubits == executor->n_qubits)) {
-        LOGGER_DEBUG("Cunqa executor already configured with {} qubits.", std::to_string(n_qubits));
-    } else {
-        executor = std::make_unique<Executor>(n_qubits);
-        LOGGER_DEBUG("Executor created with {} qubits.", std::to_string(n_qubits));
-    }
-
-    auto start_time = std::chrono::high_resolution_clock::now();
-    for (int i = 0; i < shots; i++) {
-        for (auto& instruction : instructions) {
-            instruction_name = instruction.at("name").get<std::string>();
-            qubits = instruction.at("qubits").get<std::vector<int>>();
-            switch (constants::INSTRUCTIONS_MAP.at(instruction_name))
-            {
-                case constants::MEASURE:
-                    measurement = executor->apply_measure(qubits);
-                    break;
-                case constants::UNITARY:
-                    matrix = instruction.at("params").get<Matrix>();
-                    executor->apply_unitary(matrix, qubits);
-                    break;
-                case constants::ID:
-                case constants::X:
-                case constants::Y:
-                case constants::Z:
-                case constants::H:
-                case constants::SX:
-                case constants::CX:
-                case constants::CY:
-                case constants::CZ:
-                case constants::ECR:
-                case constants::C_IF_H:
-                case constants::C_IF_X:
-                case constants::C_IF_Y:
-                case constants::C_IF_Z:
-                case constants::C_IF_CX:
-                case constants::C_IF_CY:
-                case constants::C_IF_CZ:
-                case constants::C_IF_ECR:
-                    executor->apply_gate(instruction_name, qubits);
-                    break;
-                case constants::RX:
-                case constants::RY:
-                case constants::RZ:
-                case constants::CRX:
-                case constants::CRY:
-                case constants::CRZ:
-                case constants::C_IF_RX:
-                case constants::C_IF_RY:
-                case constants::C_IF_RZ:
-                    param =  instruction.at("params").get<std::vector<double>>();
-                    executor->apply_parametric_gate(instruction_name, qubits, param);
-                    break;
-                case constants::MEASURE_AND_SEND:
-                    endpoint = instruction.at("qpus").get<std::vector<std::string>>();
-                    measurement = executor->apply_measure(qubits); 
-                    classical_channel->send_measure(measurement, endpoint[0]); 
-                    break;
-                case constants::REMOTE_C_IF_H:
-                case constants::REMOTE_C_IF_X:
-                case constants::REMOTE_C_IF_Y:
-                case constants::REMOTE_C_IF_Z:
-                case constants::REMOTE_C_IF_CX:
-                case constants::REMOTE_C_IF_CY:
-                case constants::REMOTE_C_IF_CZ:
-                case constants::REMOTE_C_IF_ECR:
-                    endpoint = instruction.at("qpus").get<std::vector<std::string>>();
-                    measurement = classical_channel->recv_measure(endpoint[0]); 
-                    if (measurement == 1) {
-                        executor->apply_gate(constants::CORRESPONDENCE_REMOTE_GATE_MAP.at(instruction_name), qubits);
-                    }
-                    break;
-                case constants::REMOTE_C_IF_RX:
-                case constants::REMOTE_C_IF_RY:
-                case constants::REMOTE_C_IF_RZ:
-                    endpoint = instruction.at("qpus").get<std::vector<std::string>>();
-                    param = instruction.at("params").get<std::vector<double>>();
-                    measurement = classical_channel->recv_measure(endpoint[0]); 
-                    if (measurement == 1) {
-                        executor->apply_gate(constants::CORRESPONDENCE_REMOTE_GATE_MAP.at(instruction_name), qubits);
-                    }
-                    break;  
-                default:
-                    LOGGER_ERROR("Invalid gate name."); 
-                    throw std::runtime_error("Invalid gate name.");
-                    break;
-            }
-        }
-        int position = executor->get_nonzero_position();
-        counts[position]++;
-        executor->restart_statevector();
-    }
-
-    auto stop_time = std::chrono::high_resolution_clock::now();
-    std::chrono::duration<float> duration = stop_time - start_time;
-    time_taken = duration.count();
-
-    result = {
-        {"counts", counts},
-        {"time_taken", time_taken}
-    }; 
-
-    counts.clear();
-
-    return result;
-
+    LOGGER_DEBUG("We are in the execute() method of CCCunqa.");
+    return cunqa_execution_<ClassicalCommBackend>(backend, quantum_task, this->classical_channel.get());
 }
 
 std::string CunqaCCSimulator::get_communication_endpoint_()
