+++ src/backends/simulators/Munich/CMakeLists.txt
@@ -1,4 +1,6 @@
 add_library(munich_simple_simulator "${CMAKE_CURRENT_SOURCE_DIR}/munich_simple_simulator.cpp")
+target_include_directories(munich_simple_simulator PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}"
+                                                            "${CMAKE_SOURCE_DIR}/src/classical_channel")
 target_link_libraries(munich_simple_simulator PRIVATE json MQT::DDSim)
