+++ src/backends/simulators/Munich/munich_classical_comm_simulator.hpp
@@ -29,8 +29,6 @@ public:
     
 
 private:
-    JSON usual_execution_(const ClassicalCommBackend& backend, const QuantumTask& quantum_task);
-    JSON distributed_execution_(const ClassicalCommBackend& backend, const QuantumTask& quantum_task);
     std::string get_communication_endpoint_() override;
 
     std::unique_ptr<comm::ClassicalChannel> classical_channel;
