+++ src/backends/simulators/Munich/munich_simple_simulator.cpp
@@ -1,3 +1,7 @@
+#include "munich_simple_simulator.hpp"
+#include "munich_executors.hpp"
+#include "munich_helpers.hpp"
+
 #include <chrono>
 #include <optional>
 
@@ -6,58 +10,20 @@
 #include "ir/QuantumComputation.hpp"
 
 #include "logger.hpp"
-#include "munich_simple_simulator.hpp"
-#include "munich_helpers.hpp"
 
 namespace cunqa {
 namespace sim {
 
 JSON MunichSimpleSimulator::execute(const SimpleBackend& backend, const QuantumTask& quantum_task)
 {
-    try {
-        // TODO: Change the format with the free functions 
-        std::string circuit = quantum_task_to_Munich(quantum_task);
-        LOGGER_DEBUG("OpenQASM circuit: {}", circuit);
-        auto mqt_circuit = std::make_unique<qc::QuantumComputation>(std::move(qc::QuantumComputation::fromQASM(circuit)));
+    LOGGER_DEBUG("We are in the execute() method of SimpleMunich.");
+    if (!quantum_task.is_dynamic) {
+        return usual_execution_<SimpleBackend>(backend, quantum_task);
+    } else {
+        return dynamic_execution_<SimpleBackend>(backend, quantum_task);
+    }   
+} 
 
-        JSON result_json;
-        JSON noise_model_json = backend.config.noise_model;
-        float time_taken;
-
-        if (!noise_model_json.empty()){
-            const ApproximationInfo approx_info{noise_model_json["step_fidelity"], noise_model_json["approx_steps"], ApproximationInfo::FidelityDriven};
-                StochasticNoiseSimulator sim(std::move(mqt_circuit), approx_info, quantum_task.config["seed"], "APD", noise_model_json["noise_prob"],
-                                            noise_model_json["noise_prob_t1"], noise_model_json["noise_prob_multi"]);
-            
-            auto start = std::chrono::high_resolution_clock::now();
-            auto result = sim.simulate(quantum_task.config["shots"]);
-            auto end = std::chrono::high_resolution_clock::now();
-            std::chrono::duration<float> duration = end - start;
-            time_taken = duration.count();
-            
-            if (!result.empty())
-                return {{"counts", JSON(result)}, {"time_taken", time_taken}};
-            throw std::runtime_error("QASM format is not correct."); 
-        } else {
-            CircuitSimulator sim(std::move(mqt_circuit));
-            
-            auto start = std::chrono::high_resolution_clock::now();
-            auto result = sim.simulate(quantum_task.config["shots"]);
-            auto end = std::chrono::high_resolution_clock::now();
-            std::chrono::duration<float> duration = end - start;
-            time_taken = duration.count();
-            
-            if (!result.empty())
-                return {{"counts", JSON(result)}, {"time_taken", time_taken}};
-            throw std::runtime_error("QASM format is not correct."); 
-        }        
-    } catch (const std::exception& e) {
-        // TODO: specify the circuit format in the docs.
-        LOGGER_ERROR("Error executing the circuit in the Munich simulator.");
-        return {{"ERROR", std::string(e.what()) + ". Try checking the format of the circuit sent and/or of the noise model."}};
-    }
-    return {};
-}
 
 } // End of sim namespace
 } // End of cunqa namespace
